+++ rotkehlchen/data_handler.py
@@ -7,6 +7,7 @@ import zlib
 from pathlib import Path
 
 from rotkehlchen.assets.asset import Asset
+from rotkehlchen.constants.misc import USERDB_NAME, USERSDIR_NAME
 from rotkehlchen.crypto import decrypt, encrypt
 from rotkehlchen.db.dbhandler import DBHandler
 from rotkehlchen.db.settings import ModifiableDBSettings
@@ -67,15 +68,15 @@ class DataHandler:
         than the one supported.
         - DBSchemaError if database schema is malformed
         """
-        user_data_dir = self.data_directory / username
+        user_data_dir = self.data_directory / USERSDIR_NAME / username
         if create_new:
             try:
-                if (user_data_dir / 'rotkehlchen.db').exists():
+                if (user_data_dir / USERDB_NAME).exists():
                     raise AuthenticationError(
                         f'User {username} already exists. User data dir: {user_data_dir}',
                     )
 
-                user_data_dir.mkdir(exist_ok=True)
+                user_data_dir.mkdir(parents=True, exist_ok=True)
             except PermissionError as e:
                 raise SystemPermissionError(
                     f'Failed to create directory for user: {e!s}',
@@ -86,7 +87,7 @@ class DataHandler:
                 if not user_data_dir.exists():
                     raise AuthenticationError(f'User {username} does not exist')
 
-                if not (user_data_dir / 'rotkehlchen.db').exists():
+                if not (user_data_dir / USERDB_NAME).exists():
                     raise PermissionError
 
             except PermissionError as e:
@@ -97,7 +98,7 @@ class DataHandler:
                 # user account can be created
                 shutil.move(
                     user_data_dir,
-                    self.data_directory / f'auto_backup_{username}_{ts_now()}',
+                    self.data_directory / USERSDIR_NAME / f'auto_backup_{username}_{ts_now()}',
                 )
 
                 raise SystemPermissionError(
@@ -169,9 +170,13 @@ class DataHandler:
         particular user is logged in or not
         """
         users = {}
-        for x in self.data_directory.iterdir():
+        users_dir = self.data_directory / USERSDIR_NAME
+        if not users_dir.exists():
+            return {}
+
+        for x in users_dir.iterdir():
             try:
-                if x.is_dir() and (x / 'rotkehlchen.db').exists():
+                if x.is_dir() and (x / USERDB_NAME).exists():
                     users[x.stem] = 'loggedin' if x.stem == self.username else 'loggedout'
             except PermissionError:
                 # ignore directories that can't be accessed
@@ -224,9 +229,10 @@ class DataHandler:
         log.info('Decompress and decrypt DB')
         # First make a backup of the DB we are about to replace
         date = timestamp_to_date(ts=ts_now(), formatstr='%Y_%m_%d_%H_%M_%S', treat_as_local=True)
+        users_dir = self.data_directory / USERSDIR_NAME
         shutil.copyfile(
-            self.data_directory / self.username / 'rotkehlchen.db',
-            self.data_directory / self.username / f'rotkehlchen_db_{date}.backup',
+            users_dir / self.username / USERDB_NAME,
+            users_dir / self.username / f'rotkehlchen_db_{date}.backup',
         )
 
         decrypted_data = decrypt(self.db.password.encode(), encrypted_data)